Weird error stuffs

+++ core/src/com/jeremyfeltracco/core/controllers/Controller.java
@@ -23,7 +23,7 @@ public abstract class Controller {
 	
 	public abstract Element generateRandomConfig();
 	
-	public abstract void mutateElement(Element e, int mutateAmt);
+	public abstract void mutateElement(Element e, float mutateAmt);
 	
 	public abstract Controller clone();
 }
